re:CR - extract render-single-menu and render-submenu-name map functions
