+++ src/domains/dashboard/components/node-view/node-view.tsx
@@ -5,62 +5,26 @@ import { useWindowScroll } from "react-use"
 
 import { name2id } from "utils/name-2-id"
 import { ChartsMetadata } from "domains/global/types"
-import { ChartDetails } from "domains/chart/chart-types"
 import { Attributes } from "domains/chart/utils/transformDataAttributes"
 
-import { renderChartsAndMenu } from "domains/dashboard/utils/render-charts-and-menu"
-import { Menu, options } from "domains/dashboard/utils/netdata-dashboard"
-import { parseChartString } from "domains/dashboard/utils/parse-chart-string"
-import { sortObjectByPriority, prioritySort } from "domains/dashboard/utils/sorting"
-import { HeadMain } from "domains/dashboard/components/head-main"
-import { MenuSidebar } from "domains/dashboard/components/menu-sidebar"
+import { renderChartsAndMenu } from "../../utils/render-charts-and-menu"
+import { Menu, options } from "../../utils/netdata-dashboard"
+import { parseChartString } from "../../utils/parse-chart-string"
+import { sortObjectByPriority, prioritySort } from "../../utils/sorting"
 
-import { ChartWrapper } from "domains/dashboard/components/chart-wrapper"
+import { HeadMain } from "../head-main"
+import { MenuSidebar } from "../menu-sidebar"
+import { ChartWrapper } from "../chart-wrapper"
+import { renderSubmenuName } from "./render-submenu-name"
+import { generateHeadCharts } from "./generate-head-charts"
 
-// needs to be imported before "dashboard_info"
-// eslint-disable-next-line import/order
-import { netdataDashboard } from "../../utils/netdata-dashboard"
+
 import "dashboard_info"
 
 import "./node-view.scss"
 
 const chartsPerRow = () => (
   options.chartsPerRow === 0 ? 1 : options.chartsPerRow
-)
-
-
-type HeadDescription = ((os: string, id: string) => string) | string
-
-function generateHeadCharts(type: string, chart: ChartDetails, duration: number) {
-  // todo don't add head charts on print view
-  // if (urlOptions.mode === 'print') {
-  //   return '';
-  // }
-
-  const hcharts = netdataDashboard.anyAttribute(netdataDashboard.context, type, chart.context, [])
-  return hcharts.map((hChart: HeadDescription) => (typeof hChart === "function"
-    ? hChart(netdataDashboard.os, chart.id)
-      .replace(/CHART_DURATION/g, duration.toString())
-      .replace(/CHART_UNIQUE_ID/g, chart.id)
-    : hChart.replace(/CHART_DURATION/g, duration.toString())
-      .replace(/CHART_UNIQUE_ID/g, chart.id)
-  ))
-}
-
-const chartCommonMin = (family: string, context: string, units: string) => (
-  netdataDashboard.anyAttribute(
-    netdataDashboard.context, "commonMin", context, undefined,
-  ) === undefined
-    ? ""
-    : `${family}/${context}/${units}`
-)
-
-const chartCommonMax = (family: string, context: string, units: string) => (
-  netdataDashboard.anyAttribute(
-    netdataDashboard.context, "commonMax", context, undefined,
-  ) === undefined
-    ? ""
-    : `${family}/${context}/${units}`
 )
 
 
@@ -107,91 +71,9 @@ const SubSection = memo(({
             )),
         )}
       </div>
-      {submenuNames.map((submenuName: string) => {
-        const submenuID = name2id(`menu_${menuName}_submenu_${submenuName}`)
-        const submenu = menu.submenus[submenuName]
-        const chartsSorted = submenu.charts
-          .concat() // shallow clone
-          .sort(prioritySort)
-        const submenuInfo = submenu.info
-        return (
-          <div
-            role="region"
-            className="dashboard-section-container"
-            id={submenuID}
-            key={submenuName}
-          >
-            <h2 id={submenuID} className="netdata-chart-alignment">
-              {submenu.title}
-            </h2>
-            {submenuInfo && (
-              <div
-                className="dashboard-submenu-info netdata-chart-alignment"
-                role="document"
-              >
-                {submenuInfo}
-              </div>
-            )}
-            <div className="netdata-chart-row">
-              {chartsSorted
-                .flatMap((chart) => generateHeadCharts("heads", chart, duration))
-                .map(parseChartString)
-                .map((attributes: Attributes | null) => (
-                  attributes && (
-                    <ChartWrapper
-                      attributes={attributes}
-                      key={`${attributes.id}-${attributes.dimensions}`}
-                    />
-                  )
-                ))}
-            </div>
-            {chartsSorted.map((chart) => {
-              const commonMin = chartCommonMin(chart.family, chart.context, chart.units)
-              const commonMax = chartCommonMax(chart.family, chart.context, chart.units)
-              return (
-                <div
-                  className="netdata-chartblock-container"
-                  style={{ width: `${pcentWidth}%` }}
-                  key={`${chart.id}-${chart.dimensions}`}
-                >
-                  {/* eslint-disable-next-line react/no-danger */}
-                  <span dangerouslySetInnerHTML={{
-                    __html: netdataDashboard.contextInfo(chart.context),
-                  }}
-                  />
-                  <ChartWrapper
-                    id={`chart_${name2id(chart.id)}`}
-                    attributes={{
-                      id: chart.id,
-                      chartLibrary: "dygraph",
-                      width: "100%",
-                      height: netdataDashboard.contextHeight(
-                        chart.context, options.chartsHeight,
-                      ),
-                      dygraphValueRange: netdataDashboard.contextValueRange(
-                        chart.context,
-                      ),
-                      before: 0,
-                      after: -duration,
-                      heightId: `${name2id(`${options.hostname}/${chart.id}`)}`,
-                      colors: `${netdataDashboard.anyAttribute(
-                        netdataDashboard.context, "colors", chart.context, "",
-                      )}`,
-                      decimalDigits: netdataDashboard.contextDecimalDigits(
-                        chart.context, -1,
-                      ),
-                      // add commonMin/commonMax attributes only if they are set
-                      ...(commonMin ? { commonMin } : {}),
-                      ...(commonMax ? { commonMax } : {}),
-                    }}
-                  />
-                </div>
-              )
-            })}
-          </div>
-        )
-      })}
-
+      {submenuNames.map(renderSubmenuName({
+        duration, menu, menuName, pcentWidth,
+      }))}
     </div>
   )
 })
